Move trades lookup fn into separate helper
