+++ workers/loc.api/sync/transaction.tax.report/helpers/index.js
@@ -1,0 +1,7 @@
+'use strict'
+
+const lookUpTrades = require('./look-up-trades')
+
+module.exports = {
+  lookUpTrades
+}
